WIP (not compiling): Obsoleting old Azure-based storage over IStorageService

+++ Apps/StorageCore/AmazonStorageServices.cs
@@ -1,11 +1,13 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Net;
 using System.Text;
 using System.Threading.Tasks;
 using Amazon;
 using Amazon.S3;
 using Amazon.S3.Model;
+using Microsoft.Azure.Storage;
 using TheBall.Core.Storage;
 
 namespace TheBall.Core.StorageCore
@@ -55,17 +57,28 @@ namespace TheBall.Core.StorageCore
         public async Task<BlobStorageItem> GetBlobItemA(IContainerOwner owner, string blobPath)
         {
             var blobAddress = GetOwnerContentLocation(owner, blobPath);
-            var response = await BlobClient.GetObjectAsync(BucketName, blobAddress);
-            var result = new BlobStorageItem(response.Key, response.ETag, response.ContentLength, response.LastModified);
+            var response = await BlobClient.GetObjectMetadataAsync(BucketName, blobAddress);
+            var result = new BlobStorageItem(blobAddress, response.ETag, response.ContentLength, response.LastModified);
             return result;
         }
 
-        public async Task DeleteBlobA(string blobPath)
+        public async Task DeleteBlobA(string blobPath, string eTag = null)
         {
+            //var blobObject = new 
+            var getObjectRequest = new GetObjectRequest()
+            {
+                BucketName = BucketName,
+                Key =  blobPath,
+                EtagToMatch = eTag,
+            };
+            var blob = await BlobClient.GetObjectAsync(getObjectRequest);
+            if (blob.HttpStatusCode == HttpStatusCode.PreconditionFailed)
+                throw new TBStorageException(HttpStatusCode.PreconditionFailed);
             await BlobClient.DeleteObjectAsync(BucketName, blobPath);
         }
 
-        public async Task<byte[]> DownloadBlobDataA(IContainerOwner owner, string blobPath, bool returnNullIfMissing)
+        public async Task<byte[]> DownloadBlobDataA(IContainerOwner owner, string blobPath, bool returnNullIfMissing,
+            string eTag = null)
         {
             var blobAddress = GetOwnerContentLocation(owner, blobPath);
             var response = await BlobClient.GetObjectAsync(BucketName, blobAddress);
@@ -76,6 +89,13 @@ namespace TheBall.Core.StorageCore
                 data = memoryStream.ToArray();
             }
             return data;
+        }
+
+        public async Task DownloadBlobStreamA(IContainerOwner owner, string blobPath, Stream stream, bool returnNullIfMissing, string eTag = null)
+        {
+            var blobAddress = GetOwnerContentLocation(owner, blobPath);
+            var response = await BlobClient.GetObjectAsync(BucketName, blobAddress);
+            await response.ResponseStream.CopyToAsync(stream);
         }
 
         public async Task<string[]> GetLocationFoldersA(IContainerOwner owner, string locationPath)
@@ -99,13 +119,14 @@ namespace TheBall.Core.StorageCore
             return result.ToArray();
         }
 
-        public async Task<BlobStorageItem> UploadBlobDataA(IContainerOwner owner, string blobPath, byte[] data)
+        public async Task<BlobStorageItem> UploadBlobDataA(IContainerOwner owner, string blobPath, byte[] data, string eTag = null)
         {
             var blobAddress = GetOwnerContentLocation(owner, blobPath);
             var req = new PutObjectRequest
             {
                 Key = blobAddress,
-                BucketName = BucketName
+                BucketName = BucketName,
+                
             };
             PutObjectResponse response;
             using (var memoryStream = new MemoryStream(data))
@@ -118,13 +139,13 @@ namespace TheBall.Core.StorageCore
             return result;
         }
 
-        public async Task<BlobStorageItem> UploadBlobTextA(IContainerOwner owner, string blobPath, string text)
+        public async Task<BlobStorageItem> UploadBlobTextA(IContainerOwner owner, string blobPath, string text, string eTag = null)
         {
             var blobAddress = GetOwnerContentLocation(owner, blobPath);
             var req = new PutObjectRequest
             {
                 Key = blobAddress,
-                BucketName = BucketName
+                BucketName = BucketName, 
             };
             PutObjectResponse response;
             using (var memoryStream = new MemoryStream())
@@ -136,6 +157,22 @@ namespace TheBall.Core.StorageCore
                 req.InputStream = memoryStream;
                 response = await BlobClient.PutObjectAsync(req);
             }
+
+            var result = await GetBlobItemA(owner, blobPath);
+            return result;
+        }
+
+        public async Task<BlobStorageItem> UploadBlobStreamA(IContainerOwner owner, string blobPath, Stream stream, string eTag = null)
+        {
+            var blobAddress = GetOwnerContentLocation(owner, blobPath);
+            var req = new PutObjectRequest
+            {
+                Key = blobAddress,
+                BucketName = BucketName,
+            };
+            PutObjectResponse response;
+            req.InputStream = stream;
+            response = await BlobClient.PutObjectAsync(req);
 
             var result = await GetBlobItemA(owner, blobPath);
             return result;
